HADOOP-17981. resilient commit.
adds options vararg similar to rename/3; allows caller to declare it is confident the
dest path doesn't exist, so no need to check for overwrites or similar.

test on local fs verifies that renaming on a filesystem without the new
interface matches that of the new behaviour,

committer test alternates between raw/local fs on parameterized runs.
minor reorg of the new tests in that class

Change-Id: Ic31d52c9a95b768e0a97583e906c20c53b120194