HADOOP-18002. abfs rename idempotency broken -remove recovery
Cut modtime-based rename recovery as it doesn't work.
Applications will have to use etag API of HADOOP-17979
and implement it themselves.

Why not do the HEAD and etag recovery in ABFS client?
Cuts the IO capacity in half so kills job commit performance.
The manifest committer of MAPREDUCE-7341 will do this recovery
and act as the reference implementation of the algorithm.

Change-Id: I071fb31967ae63e0247e2f328d9cfd0e2423b2bf