HADOOP-17409 better store downgrading; local store now recognised
if an application (such as a test suite) explicitly requested the local metadata store,
a warning is printed and the FS initialization continues without a metastore.

This is for compatibility with applications/libraries which set this to be
guaranteed a consistent store view within their test process.
It is no longer needed -and so downgraded to a no-op.

updated docs, including discussion on increased S3 IO

Change-Id: Ifacb2d9688d1852cfded17cd1ff0eb0a3200425c