HADOOP-18091. S3A Auditor Memory Leakage
This patch does not fix the problem, but it allows for auditing
to be disabled such that no ThreadLocal fields are created.
By disabling auditing by default, memory leaks will not surface.

Auditing to S3 Logs can be enabled if required; short-lived
applications and/or applications which use the same limited set of
S3A instances should be safe to use with auditing.

Change-Id: I9090d65c896a12feb6826e581935078ffb737bab