HADOOP-18091. Add regular prunings on span deactivation
If, somehow, spans were not being deactivated from threads which where
then destroyed, then the map would fill up with weak references
to those spans.

They wouldn't stop GC-ing of the audit manager, but they would fill
up the hash table.

Now, when a span is deactivated.
1. its entry is removed from the map (not replaced with the unbounded
  one, as that will happen on demand)
2. when the unbounded span is active on a thread, its deactivate()
   call also removes it from the map.
3. every so often (PRUNE_THRESHOLD = 10000), a full prune
  of the reference map is performed. this will remove any null references.

Because of 1 & 2, when valid or unbound spans are deactivated, the map shrinks.
Because of 3, even if threads are destroyed with valid spans active,
after a GC the fact that they have been GC'd will surface.

This doesn't remove map entries with unbonded spans active as the
strong references held to them means they get GC'd.

We could do that with a prune() in the audit manager removing
all entries where !isValid() is true.

Change-Id: I62fc9d66216dd8fdafb22d8b57dcdd6d311915e3