Removing reliance on SDK
Turns out last SDK upgrade broke something it shouldn't and proved how
brittle some parts of this code base are. One thing that's not brittle
is the endpoint we make requests to. Endpoint seems to be quite easy to
generate, thus we're moving away from using the SDK in a "nice" way and
we're doing it for ourselves, since it's easier to understand what's
happening (and it's 1 line of code).