MAPREDUCE-7341. feedback from sseth
* Rate limiting has been pushed into abfs fs (for the new rename only, for now).
* RateLimiter acquire() method returns a Duration instance, removing all
  ambiguity about unit of measurement.

This cuts it out of all of the manifest committer code, config etc.

It is still useful to collect stats on how often rate limiting is occurring,
so the store operation's commitFile() method includes that duration value
in its CommitFileResult result tuple. This value makes its way
into the _SUCCESS report.

To simplify binding to abfs, a new committer factory has been added to
the hadoop-azure module:
org.apache.hadoop.fs.azurebfs.commit.AzureManifestCommitterFactory

This is a subclass of ManifestCommitterFactory which patches the configuration
to always use the AbfsManifestStoreOperations store operations class
to commit work.

Also

* marked PathOutputCommitterFactory as public/evolving.
It has been out for a few years, used in spark etc and we have to be
careful about changing it.
* written more on the algorithm itself.

Change-Id: I7e624cacb9106dc4f5e6e8b5ef5c646cc3d734c0