HADOOP-15980 : Enable TLS in RPC client/server
Adding support for reading ssl configuration from ssl-server.xml and
ssl-client.xml. This unifies this patch with the rest of the hadoop SSL
ecosystem.

The patch has added a file SSLHandlerProvider that reads the ssl
configuration files and creates the netty SslHandler object. This can be
added to a Netty pipeline to encrypt the communication through the
pipeline.

The patch also adds a test case for the SSLHandlerProvider.

In addition to this the patch modifies TestNettyRPC to read the SSL
configuration from ssl-server.xml and ssl-client.xml.