MAPREDUCE-7341. Optimise task commit directory probes
we now execute the getFileStatus call after listing the source dir and
reading the first result of that listing.
If the store performs async listings -which abfs does- the HEAD will
overlap with the LIST being executed in a separate thread

we also track if the parent dest dir exists...if it doesn't there
is no need to look for child destination directories.

Needs a test of cost of this, really.

Change-Id: I65ba26e2484df96fe2d45c15f67692c7c4c1a89d