HADOOP-16202. s3a openfile to support a new drain policy fs.s3a.input.async.drain.threshold
If the #of bytes is above this, the stream is drained in a separate thread;
the threshold is there because for small amounts, scheduling the work seems
to be slower than the actual processing.

draining is also done via a 16K buffer, which reduces a lot of OS API calls.

Change-Id: I1b2a71a05c37ada289cf23e128da0a6b01452ee6