HADOOP-18202. create-release fails fatal: unsafe repository
Since CVE-2022-24765 in April 2022, git refuses to work in directories
whose owner != the current user, unless explicitly told to trust it.

This patches the create-release script to trust the /build/source
dir mounted from the hosting OS, whose userid is inevitably different
from that of the account in the container running git.

Change-Id: I1646d7f5d098252c12b000c51a99f9df11919306