HADOOP-17833. Tuning commit performance
Marginal gains in task commit

* Always say there is data to commit
* skip task cleanup and its list/delete operation, delaying until
  job cleanup

Delaying task attempt cleanup is an interesting one.
it saves the TA LIST, smaller DELETE batch and the probes for
recreating dirs (which aren't being skipped (yet))

but it does postpone to the job commit.
If the # of files/dir markers across a job is less than a few thousand,
the existing single list call + overlapped delete batch will absorb the
cost.
If it is very large, there's a risk we move cleanup here.
maybe: when loading each single file, delete that and the
stub file in the parallel phase. no need to list, one single
delete to cover both. job commit cleanup is then only to list
dir markers

Change-Id: Ia91d643668454c23cd1df2e5abb40cad806b671c