HADOOP-17833. round out createFile() with header options
Wrapping up createFile() with the ability to set headers on the file

builder.must("fs.s3a.create.header.my-header","my-value");

this will set a user metadata entry "my-header" which xattr will return
as "header.my-header".

This forced me to make PutOptions something passed around more rigorously and
used as the parameter class to innerCreateFile. This is ultimately good as it
lets us set new options later..encryption, storage class etc. if we so choose.

also: s3a path capabilities let you probe for this prefix, as well as the
performance one...think this should be the standard behaviour for custom options.

Change-Id: Icb5a1e923a3bc50f42e825d34b16da0cf622c289