HADOOP-18410. copy fields to variables before use in a lambda expression
the key cause is that even though the fields passed in to drain() were
converted to references through the methods, in the lambda expression
passed in to submit, they are direct references
```
operation = client.submit(
 () -> drain(uri, streamStatistics,
       false, reason, remaining,
       object, wrappedStream));  /* here */
```
the fields are only read during the async execution, not during the submit phase.

the next step the code does is reset those fields; the async work is
failing with an NPE which isn't being noted.

Adding the join() after submit works *because this was inserted before
the fields were set to null...so no NPE being silently raised.

Change-Id: I417108db0daa78f711eb8dc1ed96ef619437a2bf