HDFS-16740. Mini cluster test flakiness
Add a constructor to the mini cluster constructors and builders that takes a TemporaryFolder.  This encourages the use of managed temporary folders that are unique to the test and automatically cleaned up.  The field is named "baseDir" to communicate that this folder acts as the base directory for test data.

These subset of tests recently demonstrated flaky behavior for multiple pull requests.  Switch to using a temporary folder managed by the test class, and ensure that the test cluster is shutdown using either @After or try-with-resources.  Using the new constructor encourages the consistent use of the managed temporary folder.

Allow the same baseDir to be provided multiple times, simplifying cluster restarts.  This simplifies these tests as the use of the same TemporaryFolder will always work (even for restarts)

This change also reduces the number of tests that make assumptions about the  mini clusters using encapsulation.  These changes avoid configurations that directly use HDFS_MINIDFS_BASEDIR.