HADOOP-18526. Leak of S3AInstrumentation instances via hadoop Metrics references
Add tests that instrumentation.close() unregisters *once*.

reverted the change to close() it in s3afs.close(), because that was
already done in stopAllServices().

if something is leaking S3AInstrumentation instances, it is because
it is creating fs instances outside the cache, and not closing them.
once dereferenced and a GC triggered, the fs instance and most
internal classes are GC'd, but not the instrumentation because of
the back ref.

This weakref link will do that, but will result in the metrics structure
being filled up with lots of weakrefs which don't get cleaned up.
It's a lot less expensive than strong refs, but still not ideal.
Fixing that would be tricky as the base metrics class or a subclass
would need to do cleanups (when?).

Change-Id: I07fe792301e23ed1d3676fbb8cc033c5a96e55a2