HADOOP-18456 probes feedback
Use "fs.capability.etags.available" as the probe name.

Had to create a new file for this as none of the existing
ones were appropriate.
org.apache.hadoop.fs.azurebfs.constants.InternalConstants

Print this value in the toString() of abfs and AbfsInputStream.

Tests for this, including a new one in ITestFileSystemInitialization
which includes tests for other expected capabilities.

Change-Id: I31cbb58ef23ce78b32572c7f86a8acf269040e55