HDFS-16906. Fixed leak in CryptoOutputStream::close
Fix a bug in the close() method of CryptoOutputStream.

When closing we need to wrap the flush() in a try .. finally, otherwise
when flush throws it will prevent us completing the remainder of the
close activities and in particular the close of the underlying wrapped
stream object resulting in a resource leak.