HADOOP-18636 LocalDirAllocator cannot recover from directory tree deletion
Even though DiskChecker.mkdirsWithExistsCheck() will create the directory tree,
it is only called *after* the enumeration of directories with available
space has completed.
Directories which don't exist are reported as having 0 space, therefore
the mkdirs code is never reached.
Adding a simple mkdirs() -without bothering to check the outcome-
ensures that if a dir has been deleted then it will be reconstructed
if possible. If it can't it will still have 0 bytes of space
reported and so be excluded from the allocation.

No need to call exists() first as that is what mkdirs() does first anyway.

Change-Id: I8a8db7aafa64d39bb93c659f9279f3a90930eb79