HDFS-17000. Fix faulty loop condition in TestDFSStripedOutputStreamUpdatePipeline
The loop iterates over an int but previously compared the value against
Long.MAX_VALUE, so it would have overflown before having reached that
value, and the loop would never have terminated normally.

However, as mentioned in the comments of HDFS-17000 eventually an
(expected) exception would have occurred so the loop would have
terminated at some point.

Either way, this commit fixes the loop condition to be on the safe side.