Merged PR 4124487: Add config to startup NameNode as Observer
#Why
NameNode currently uses a "StartupOption" to decide whether to start a node as an Observer NameNode. This causes an issue during a rolling upgrade because "rollingUpgrade" is also a StartupOption and NameNode will only allow 1 startup option, choosing the last startup option in the list. Observer starts with the following startup options: ["-rollingUpgrade", "started", "-observer"]. This means that the rolling upgrade gets ignored which causes Observer to have an issue when an actual rolling upgrade is ongoing:
```
2023-08-23T14:59:03.486-0700,WARN,[main],org.apache.hadoop.hdfs.server.namenode.FSNamesystem,"Encountered exception loading fsimage
java.io.IOException:
File system image contains an old layout version -63.
An upgrade to version -64 is required.
Please restart NameNode with the \'-rollingUpgrade started\' option if a rolling upgrade is already started; or restart NameNode with the \'-upgrade\' option to start a new upgrade.
	at org.apache.hadoop.hdfs.server.namenode.FSImage.recoverTransitionRead(FSImage.java:271)
	at org.apache.hadoop.hdfs.server.namenode.FSNamesystem.loadFSImage(FSNamesystem.java:1116)
	at org.apache.hadoop.hdfs.server.namenode.FSNamesystem.loadFromDisk(FSNamesystem.java:724)
	at org.apache.hadoop.hdfs.server.namenode.NameNode.loadNamesystem(NameNode.java:681)
	at org.apache.hadoop.hdfs.server.namenode.NameNode.initialize(NameNode.java:763)
	at org.apache.hadoop.hdfs.server.namenode.NameNode.<init>(NameNode.java:1013)
	at org.apache.hadoop.hdfs.server.namenode.NameNode.<init>(NameNode.java:992)
	at org.apache.hadoop.hdfs.server.namenode.NameNode.createNameNode(NameNode.java:1743)
	at org.apache.hadoop.hdfs.server.namenode.NameNode.main(NameNode.java:1811)
"
```

This is needed so that we can smoothly upgrade from Hadoop 2.9 MT to Hadoop 3.3 MT. It is also necessary so that we can rollback the fsimage conversion change in this PR smoothly: https://msasg.visualstudio.com/Multi%20Tenancy/_git/Hadoop/pullrequest/4091640

#What
Add config to startup NameNode as Observer.

This change is also needed in branch-3.3-mt

#Test
Tested in MTPrime-Test-MWHE01 where observer couldn't start during a rolling upgrade due to the version mismatch exception above. After the change observer was able to start with rolling upgrade logic and observer state.