HADOOP-18679. Add API for bulk/paged object deletion
Initial pass at writing an API for bulk deletes,
targeting S3 and any store with paged delete support.

Minimal design of a RemoteIterator to provide the list
of paths to delete; a progress report will be provided
after pages are deleted so as to provide an update
of files deleted, and a way for the application
code to abort an ongoing delete -such as after a failure.

Change-Id: I3dcbb144232d76b5d4ebf7ad080d187edd6e93e4