YARN-11578. Cache fs supports chmod in LogAggregationFileController.
The check introduced in YARN-10901 to avoid a warn message in NN logs in
certain situations (when /tmp/logs is not owned by the yarn user),
but it adds 3 NameNode calls (create, setpermission, delete) during
log aggregation collection, for every NM.
Meaning, when a YARN job completes, at the YARN log aggregation
phase this check is done for every job, from every NodeManager.

In 30 minutes 4.2 % of all the NameNode calls were due to this in
a cluster. "write" calls need a Namesystem writeLock as well,
so the impact is bigger.

Change-Id: I65468aa972860d3b62050fcb41b8b06e417ee8bb