HADOOP-18925. S3A: option to enable/disable CopyFromLocalOperation
Add a new option
fs.s3a.optimized.copy.from.local.enabled

This will enable (default) or disable the
optimized CopyFromLocalOperation upload operation
when copyFromLocalFile() is invoked.

When false the superclass implementation is used; duration
statistics are still collected, though audit span entries
in logs will be for the individual fs operations, not the
overall operation.

ITestS3ACopyFromLocalFile parameterized to run with/without
option.

This is for debugging.

Change-Id: I208b8babd9c9951acbd603d811b4f5f94620892d