HADOOP-19027. S3A: S3AInputStream doesn't recover from HTTP/channel exceptions
If a read() on the inner wrapped stream returned -1 then it is closed.

There is no attempt to actually recover within a readFully(); there's
a switch to turn this on, but if anyone does it a test will spin forever
as the inner PositionedReadable.read(position, buffer, len) downgrades
all EOF exceptions to -1.
A new method would need to be added which controls whether to downgrade/rethrow
exceptions, which makes for more complex work.

What does that mean? Possibly reduced resilience to non-retried failures
on the inner stream, even though more channel exceptions are retried on.

+Split out tests of different read methods into their own test cases.

Change-Id: I8f7e75ba73a376e7dfd06c5bb9ebc4138bc80394