HADOOP-19045. CreateSession Timeout
New test ITestCreateSessionTimeout to verify that the duration set
in fs.s3a.connection.request.timeout is passed all the way down.

This is done by adding a sleep() in a custom signer and verifying
that it is interrupted and that an AWSApiCallTimeoutException is
raised.

Improves javadocs of fs.s3a.connection.request.timeout.

Change-Id: I22ee270c536fe5d0a509b707e4cf0202cc930bb5