HADOOP-18679. Bulk Delete: initial S3AStore interface/impl
Using this PR to start with the initial design, implementation
and services offered by having lower-level interaction with S3
pushed down into an S3AStore class, with interface/impl split.

The bulk delete callbacks now to talk to the store, *not* s3afs,
with some minor changes in behaviour (IllegalArgumentException is
raised if root paths / are to be deleted)

Mock tests are failing; I expected that: they are always brittle.

What next? get this in and then move lower level fs ops
over a method calling s3client at a time, or in groups, as appropriate.

The metric of success are:
* all callback classes created in S3A FS can work through the store
* no s3client direct invocation in S3AFS

Change-Id: Ib5bc58991533fd5d13e78426e88b884b6ae5205c