MAPREDUCE-7474. use commitFile() to rename the temporary task manifest
Move from classic rename() to commitFile() to rename the file,
after calling getFileStatus() to get its length and possibly etag.

This becomes a rename() on gcs/hdfs anyway, but on abfs it does reach
the ResilientCommitByRename callbacks in abfs, which report on
the outcome to the caller...which is then logged at WARN.

test changes to match the codepath changes, including improvements
in fault injection.

Change-Id: I757a77c8d2b7a7f1cf2ce32d109ce1baa6a90ec2