HADOOP-19205. pull out lazy construction classes
add class LazyAtomicReference<T>

This take a constructor CallableRaisingIOE which is then
invoked on demand to create the reference value the first time
it is evaluated in get().

It has a getUnchecked() to wrap any IOEs raised,
and implements CallableRaisingIOE<T> so its get()
call can be chained/wrapped etc.

The subclass LazyAutoCloseableReference<T extends AutoCloseable>
takes AutoCloseables, implements the interface and in close() will close
the inner reference if it is set.

These are used in ClientManagerImpl instead of doing all the null checks &c
itself.

Tests for parallel creation are now more sophisticated
* semaphore + sleep choreography
* assertions on thread ID of thread creating the class
This should be a lot less brittle to timing issues

Change-Id: Idf9770c072b0436331a76f983e533528932dceff